E-Commerce Sales Dashboard Project
Objective
An Online E-Commerce Company wants to design a Sales dashboard to analyze the sales based on various product categories. The company wants to add user control for product category, so users can select a category and can see the trend month-wise and product-wise accordingly.  
Prepare a table of Sales and Profit month-wise in one sheet, named it as ‘Working Sheet’.    
Prepare the sales table region-wise in the working sheet.    
Create a User Control Combo box for Product Category.
Create Column Chart of a month-wise table and region-wise table
Link the table with combo box.    
Create a dashboard.